+++ fs/index.ts
@@ -1,5 +1,0 @@
-import files from "./files";
-
-export default {
-	...files,
-}